Split logic by creating TransactionGroupListExpanded
