+++ src/components/SelectionListWithSections/types.ts
@@ -25,7 +25,7 @@ import type SpendCategorySelectorListItem from '@pages/workspace/categories/Spen
 // eslint-disable-next-line no-restricted-imports
 import type CursorStyles from '@styles/utils/cursor/types';
 import type CONST from '@src/CONST';
-import type {PersonalDetailsList, Policy, Report, TransactionViolation, TransactionViolations} from '@src/types/onyx';
+import type {PersonalDetailsList, Policy, Report, SearchResults, TransactionViolation, TransactionViolations} from '@src/types/onyx';
 import type {Attendee, SplitExpense} from '@src/types/onyx/IOU';
 import type {Errors, Icon, PendingAction} from '@src/types/onyx/OnyxCommon';
 import type {
@@ -520,6 +520,22 @@ type TransactionGroupListItemProps<TItem extends ListItem> = ListItemProps<TItem
     violations?: Record<string, TransactionViolations | undefined> | undefined;
 };
 
+type TransactionGroupListExpandedProps<TItem extends ListItem> = Pick<
+    TransactionGroupListItemProps<TItem>,
+    'showTooltip' | 'canSelectMultiple' | 'onCheckboxPress' | 'columns' | 'groupBy' | 'accountID' | 'isOffline' | 'violations' | 'areAllOptionalColumnsHidden'
+> & {
+    transactions: TransactionListItemType[];
+    transactionsVisibleLimit: number;
+    setTransactionsVisibleLimit: React.Dispatch<React.SetStateAction<number>>;
+    isEmpty: boolean;
+    isGroupByReports: boolean;
+    transactionsSnapshot?: SearchResults;
+    shouldDisplayEmptyView: boolean;
+    transactionsQueryJSON?: SearchQueryJSON;
+    isInSingleTransactionReport: boolean;
+    searchTransactions: (pageSize?: number) => void;
+};
+
 type ChatListItemProps<TItem extends ListItem> = ListItemProps<TItem> & {
     queryJSONHash?: number;
 
@@ -970,6 +986,7 @@ export type {
     SingleSelectListItemProps,
     MultiSelectListItemProps,
     TransactionGroupListItemProps,
+    TransactionGroupListExpandedProps,
     TransactionGroupListItemType,
     TransactionReportGroupListItemType,
     TransactionMemberGroupListItemType,
